+++ client/src/components/signup/SignupForms/index.tsx
@@ -1,8 +1,8 @@
 import SignupTerms from '../SignupTerms';
-import FirstSignupForm from '../FirstSignupForm';
+import SignupBasicForm from '../SignupBasicForm';
 import SocialLogin from 'components/login/SocialLogin';
 import { Link } from 'react-router-dom';
-import SecondSignupForm from '../SecondSignupForm';
+import SignupDetailForm from '../SignupDetailForm';
 import styles from './SignupFormController.module.scss';
 import classNames from 'classnames/bind';
 
@@ -19,7 +19,7 @@ const SignupForms = ({ step, totalSteps }: SignupFormsProps) => (
       <SignupTerms />
     </div>
     <div className={cn('page', { expand: step === 2 })}>
-      <FirstSignupForm />
+      <SignupBasicForm />
       <div className={cn('link')}>
         <Link to="/login" className={cn('login')}>
           이미 계정이 있으신가요?
@@ -28,7 +28,7 @@ const SignupForms = ({ step, totalSteps }: SignupFormsProps) => (
       <SocialLogin />
     </div>
     <div className={cn('page', { expand: step === totalSteps })}>
-      <SecondSignupForm />
+      <SignupDetailForm />
     </div>
   </>
 );
